add tag model, modify minors
