+++ django_app/recipe/models/__init__.py
@@ -1,2 +1,2 @@
 from .recipe import *
-from .recipe_ingredient import *
+from .recipe_extra import *
